Move the logic for indentation of first line of file
This moves the logic of indentation of the first line of the file from
the logic of indenting the first line of a block of any kind. This
greatly simplify the code in the block function and it doesn't make
sense to do that check that late either since it's a simple check that
hardly can go wrong.