Don't freeze full executable paths at load time
It's bad practice to use `executable-find` in defcustoms, because the
value will be expanded into a fully-qualified path at the time the
file is loaded, using whatever is the effective `exec-path` at the
time. This can lead to confusion if the file is somehow loaded before
a user deliberately sets up his path.

When only an executable name is used, as is the case with this change,
it is looked up in the $PATH at the time it is used, which reliably
leads to expected behaviour. And the user can of course still
customise any `nix-*-executable` variable to a full path if they wish.