Split up `nix-search` into two separate functions
The main intention behind this is that patching the functions should be
easier. It also allows for experimenting with different displaying
methods for the search result.

The second argument to the `display` function has been removed in order
to not dictate how the buffer should be displayed. Users can customize
this with `display-buffer-alist`.

Also, the fact that the display function is separate allows, for
example, to store a search result somewhere and have it display later.