nix-flake: Follow the terminology and syntax of the guideline
According to the guideline, *arguments* refer to required arguments,
which is either a flake ref or installable for most commands.
*Options* refer to both boolean flags and other optional arguments
starting with hyphen. It is better to stick with the terminology.

Also, options are appended to the command line after the required
arguments, so follow the rule as well.

For details, see https://nixos.org/manual/nix/stable/contributing/cli-guideline.html